updated domain objects
