+++ src/main/java/edu/gmu/cs321/ImmRelative.java
@@ -1,27 +1,31 @@
 package edu.gmu.cs321;
 
 public class ImmRelative extends Immigrant {
-    private int alienRegistrationNumber;
+    private String alienRegistrationNumber;
     private String relationship;
     private String nationality;
+    private String citizenshipStatus;
 
     public ImmRelative(Builder builder) {
         super(builder);
         this.alienRegistrationNumber = builder.alienRegistrationNumber;
         this.relationship = builder.relationship;
         this.nationality = builder.nationality;
+        this.citizenshipStatus = builder.citizenshipStatus;
     }
 
-    public int getAlienRegistrationNumber() { return alienRegistrationNumber; }
+    public String getAlienRegistrationNumber() { return alienRegistrationNumber; }
     public String getRelationship() { return relationship; }
     public String getNationality() { return nationality; }
+    public String getCitizenshipStatus() { return citizenshipStatus; }
 
     public static class Builder<T extends Builder <T>> extends Immigrant.Builder<T> {
-        protected int alienRegistrationNumber;
+        protected String alienRegistrationNumber;
         protected String relationship;
         protected String nationality;
+        protected String citizenshipStatus;
 
-        public T setAlienRegistrationNumber(int alienRegistrationNumber) {
+        public T setAlienRegistrationNumber(String alienRegistrationNumber) {
             this.alienRegistrationNumber = alienRegistrationNumber;
             return self();
         }
@@ -33,6 +37,11 @@ public class ImmRelative extends Immigrant {
 
         public T setNationality(String nationality) {
             this.nationality = nationality;
+            return self();
+        }
+
+        public T setCitizenshipStatus(String citizenshipStatus) {
+            this.citizenshipStatus = citizenshipStatus;
             return self();
         }
