+++ src/main/java/edu/gmu/cs321/Immigrant.java
@@ -1,32 +1,23 @@
 package edu.gmu.cs321;
 
 public class Immigrant extends Person {
-    private int immID;
-    private String citizenshipStatus;
+    private String ssn;
 
     protected Immigrant(Builder builder) {
         super(builder);
-        this.immID = builder.immID;
-        this.citizenshipStatus = builder.citizenshipStatus;
+        this.ssn = builder.ssn;
     }
 
-    public int getID() { return immID; }
-    public String getCitizenshipStatus() { return citizenshipStatus; }
+    public String getSSN() { return ssn; }
 
     public static class Builder<T extends Builder<T>> extends Person.Builder<T> {
-        protected int immID;
-        protected String citizenshipStatus;
+        protected String ssn;
 
-        public T setID(int immID) {
-            this.immID = immID;
+        public T setSSN(String ssn) {
+            this.ssn = ssn;
             return self();
         }
-
-        public T setCitizenshipStatus(String citizenshipStatus) {
-            this.citizenshipStatus = citizenshipStatus;
-            return self();
-        }
-
+        
         @Override
         public Immigrant build() { return new Immigrant(this); }
     }
